lantern revocations page generalized
